Create README.md
IT management student, passionate about coding.
Learning and making progress in this field are my everyday habit...❤️
I consider myself an ambitious and a collaborative person and when I accept a challenge I take it to the end, I won’t stop until I have accomplished my objectives. I have the need to feel that I gave my best and that I performed the best job possible with the resources available to do so. 
On a more personal level, I like to read and to gather new knowledge about how everything around me works...